Helios: vérification des CRL lors de la signature

+++ lib/XadesSignature.class.php
@@ -201,20 +201,25 @@ class XadesSignature {
 				return false;
 			}
 
-			continue;
-
 			$certif = strval($signatureNode->children(self::NS_DS_URI)->KeyInfo->X509Data->X509Certificate);
 
 			$content = "-----BEGIN CERTIFICATE-----\n".$certif."\n-----END CERTIFICATE-----\n";
 			$file = "/tmp/s2low_xades_".mt_rand(0,getrandmax());
 			file_put_contents($file,$content);
 
+			$command = OPENSSL_PATH . " x509 -hash -noout -in " . $file;
+			exec($command,$output,$return_var);
+
+			$file = $this->validca_path."/{$output[0]}.r0";
+
+			if (! file_exists($file)){
+				continue;
+			}
 
 			$command = OPENSSL_PATH." verify -CApath ".$this->validca_path." -crl_check $file ";
 			exec($command,$output,$return_var);
 
 			$this->last_output = implode("\n",$output);
-
 			unlink($file);
 
 			if ($return_var != 0){
